+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Applet uid & version can be specified as environment variables `SOS_APPLET_UID` & `SOS_APPLET_VERSION`.
+- Command `sos applet upload` optionally accepts `--no-update-package-config` argument which prevents updating package.json config.
 
 ## [0.9.3] - 2021-10-20
 ### Fixed